Use python-fzf in-place of pyfzf
This one has no dependencies while pyfzf depends on plumbum (that itself
depends on multiple other libraries). Also it has a better API.

The only issue is that python-fzf depends on Python 3.9.2 at minimum, so
let's change the default Python to version 3.9.

Good thing this is Nix, so this is not as much as problem as it would
be.