Remove i686-linux from supported systems
It is not evaluating in i686-linux because some strange issue with
pandas (why it is trying to evaluate it?). I could try to fix, but
i686-linux is not a supported NixOS system anyway so it is easier to
remove it from the Flake.

Folks that still want to use this can use the overlay.