Remove almost all functions from __all__
I just broke the API, I don't really want to maintain a stable API
and I also don't think anyone is using this as a library (it is not even
published somewhere interesting like PyPi).

So let's remove ilusion that we have a public API. If you still want to
do so you can of course, but this at your risk now.